ai_resume_analyzer/requirements.txt
Django
PyPDF2
python-docx
ollama